serialpps: split polling tests by source file
Polling tests and their fake hardware helpers had accumulated in serialpps_test.go alongside detection, generator, and statistics coverage.\n\nMove them to poll_test.go so the tests follow the source layout and provide a natural home for polling controller simulations. Test behavior is unchanged.
